requirements.txt
streamlit==1.20.0
numpy==1.21.2
pydicom==2.3.1
ultralytics==8.0.145
python-magic==0.4.27
SimpleITK==2.2.1
